requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-core
langchain-google-genai
langgraph
